refactor: extract dynamics decorators to submodule
